Moved location of color classes
And preserve B/C
